Revise and simplify when/default semantics.
This matches behavior seen in implementation and relied upon in the
wild. There are various examples of existing code making use of the
fact every block has an implicit C<$_>, which it sets so it can use
C<when> and C<default>. This means trying to identify a topicalizer
in a clever way, or even complain in the absence of one, breaks a
common and useful pattern (yes, I tried it!) Added some notes on
the semantics of nested when/default, also bringing the design doc
in line with existing and tested behavior.