remove an Exception To The Rule about blocks in method calls
Bare blocks and pointies are now normal arguments in a method call
introduced by a colon. Their right curly will still terminate a
statement when a newline is following, but a method call attached
to the block is now called on the block itself, rather than on the
term on the left of the block.
It is easier to add parens to get the other meaning, but it is hard
to remove implicit parens as the design docs status up to now.